Trap and melee enemy complete
Two types of trap are created, one stay at the same place and one can move around regularly.

Melee enemy was created, it can move between the leftedge and rightedge.

sprite can be added to those prefabs, and if the animation is needed, please let me know and I can add the animation.